# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v0/invoices": {
      "post": {
        "operationId": "InvoicesController_emit",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvoiceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "status ISSUED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponseModel"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (details lists the field messages)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Logto token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Trial expired (only on routes not allowed after expiry)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Company not registered yet, or resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Emissão concorrente; repetir",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Sem certificado / vencido, ou nota rejeitada pelo fisco (details.code, details.reason; nota salva como REJECTED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "API Nacional indisponível; nota fica PENDING",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "logto": []
          }
        ],
        "summary": "Emite uma NFS-e (síncrono). customerId OU tomador*; saveCustomer só com tomador*",
        "tags": [
          "invoices"
        ]
      },
      "get": {
        "operationId": "InvoicesController_findAll",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "ISSUED",
                "REJECTED",
                "CANCELLED"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PageMeta"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/InvoiceResponseModel"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (details lists the field messages)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Logto token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Trial expired (only on routes not allowed after expiry)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Company not registered yet, or resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "logto": []
          }
        ],
        "summary": "Lista notas (mais recentes primeiro)",
        "tags": [
          "invoices"
        ]
      }
    },
    "/api/v0/invoices/{id}": {
      "get": {
        "operationId": "InvoicesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponseModel"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (details lists the field messages)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Logto token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Trial expired (only on routes not allowed after expiry)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Company not registered yet, or resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "logto": []
          }
        ],
        "summary": "Detalhe da nota",
        "tags": [
          "invoices"
        ]
      }
    },
    "/api/v0/invoices/{id}/xml": {
      "get": {
        "operationId": "InvoicesController_getXml",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/xml": {
                "schema": {
                  "type": "string",
                  "format": "xml"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (details lists the field messages)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Logto token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Trial expired (only on routes not allowed after expiry)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Company not registered yet, or resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "logto": []
          }
        ],
        "summary": "XML da NFS-e (só ISSUED/CANCELLED)",
        "tags": [
          "invoices"
        ]
      }
    },
    "/api/v0/invoices/{id}/pdf": {
      "get": {
        "operationId": "InvoicesController_getPdf",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (details lists the field messages)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Logto token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Trial expired (only on routes not allowed after expiry)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Company not registered yet, or resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "logto": []
          }
        ],
        "summary": "DANFSe em PDF (Content-Disposition: inline)",
        "tags": [
          "invoices"
        ]
      }
    },
    "/api/v0/invoices/{id}/cancel": {
      "post": {
        "operationId": "InvoicesController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelInvoiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "status CANCELLED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponseModel"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (details lists the field messages)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Logto token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Trial expired (only on routes not allowed after expiry)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Company not registered yet, or resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Nota não está ISSUED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Fisco recusou o cancelamento",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "logto": []
          }
        ],
        "summary": "Cancela uma nota ISSUED (motivo 15–255 chars)",
        "tags": [
          "invoices"
        ]
      }
    },
...
  },
  "info": {
    "title": "MATHEO NFS-e API",
    "description": "Emissor de NFS-e para MEI. Todas as rotas exigem Bearer token do Logto (audience https://api.matheo.local), exceto /health e esta documentação. Erros seguem { statusCode, message, details?, requestId }. Ver docs/frontend/api-brief.md para fluxos e estados.",
    "version": "0.3.2",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "logto": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 422
          },
          "message": {
            "type": "string",
            "example": "NFS-e rejected by the national API"
          },
          "details": {
            "description": "Validation: array of field messages. Business errors: object with context (e.g. { invoiceId, code, reason }).",
            "oneOf": [
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              {
                "type": "object",
                "additionalProperties": true
              }
            ]
          },
          "requestId": {
            "type": "string",
            "example": "a1b2c3d4-e5f6-4a7b-8c9d-0e1f2a3b4c5d",
            "description": "Same value as the x-request-id response header"
          }
        },
        "required": [
          "statusCode",
          "message",
          "requestId"
        ]
      },
...
      "PageMeta": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          },
          "total": {
            "type": "number",
            "example": 37
          }
        },
        "required": [
          "page",
          "limit",
          "total"
        ]
      },
...
      "CreateInvoiceDto": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string",
            "format": "uuid",
            "description": "Cliente salvo. Exclusivo com tomadorDocumento/tomadorNome"
          },
          "tomadorDocumento": {
            "type": "string",
            "example": "12345678909",
            "description": "Obrigatório sem customerId. CPF (11) ou CNPJ (14)"
          },
          "tomadorNome": {
            "type": "string",
            "example": "Cliente Exemplo",
            "description": "Obrigatório sem customerId"
          },
          "tomadorEmail": {
            "type": "string",
            "example": "cliente@x.com"
          },
          "saveCustomer": {
            "type": "boolean",
            "default": false,
            "description": "Só com tomador*: salva o tomador como cliente da empresa"
          },
          "descricao": {
            "type": "string",
            "example": "Consultoria em tecnologia da informação",
            "maxLength": 2000
          },
          "valor": {
            "type": "number",
            "example": 150,
            "description": "Valor do serviço, até 2 casas decimais"
          },
          "codigoTributacao": {
            "type": "string",
            "example": "01.01.01",
            "description": "Código nacional de tributação (NN.NN.NN)"
          }
        },
        "required": [
          "descricao",
          "valor",
          "codigoTributacao"
        ]
      },
      "InvoiceResponseModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "companyId": {
            "type": "string",
            "format": "uuid"
          },
          "customerId": {
            "type": "string",
            "nullable": true,
            "format": "uuid",
            "description": "Saved customer used on emission, if any"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "ISSUED",
              "REJECTED",
              "CANCELLED"
            ]
          },
          "dpsNumero": {
            "type": "number",
            "example": 7,
            "description": "Sequential DPS number per company/series"
          },
          "dpsSerie": {
            "type": "string",
            "example": "1"
          },
          "tomadorDocumento": {
            "type": "string",
            "example": "12345678909"
          },
          "tomadorNome": {
            "type": "string",
            "example": "Cliente Exemplo"
          },
          "tomadorEmail": {
            "type": "string",
            "nullable": true
          },
          "descricao": {
            "type": "string",
            "example": "Consultoria em TI"
          },
          "valor": {
            "type": "string",
            "example": "150.00",
            "description": "Decimal string with 2 places"
          },
          "codigoTributacao": {
            "type": "string",
            "example": "01.01.01"
          },
          "chaveAcesso": {
            "type": "string",
            "nullable": true,
            "example": "35503082...",
            "description": "50-digit access key once ISSUED"
          },
          "numeroNfse": {
            "type": "string",
            "nullable": true,
            "example": "7"
          },
          "rejectionReason": {
            "type": "string",
            "nullable": true,
            "description": "\"code: reason\" from the national API when REJECTED"
          },
          "cancelledAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "cancelReason": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "companyId",
          "customerId",
          "status",
          "dpsNumero",
          "dpsSerie",
          "tomadorDocumento",
          "tomadorNome",
          "tomadorEmail",
          "descricao",
          "valor",
          "codigoTributacao",
          "chaveAcesso",
          "numeroNfse",
          "rejectionReason",
          "cancelledAt",
          "cancelReason",
          "createdAt",
          "updatedAt"
        ]
      },
      "CancelInvoiceDto": {
        "type": "object",
        "properties": {
          "motivo": {
            "type": "string",
            "example": "Erro de digitação no valor do serviço",
            "minLength": 15,
            "maxLength": 255
          }
        },
        "required": [
          "motivo"
        ]
      },
...
    }
  }
}